Second version - better style, add basic form validation

+++ src/components/ShowFood.vue
@@ -1,12 +1,20 @@
 <template>
   <div class="card">
-    <h2>Added food</h2>
     <!-- <button @click="loadData">Load data</button> -->
     <ul>
-      <li :class="{alert: food.difference <=2 && food.difference >0, expired: food.difference<0}" v-for="food in foods" :key="food.id">
-        <p>{{ food.name }}</p>
+      <li
+        :class="{
+          alert: food.difference <= 2 && food.difference > 0,
+          expired: food.difference < 0,
+        }"
+        v-for="food in foods"
+        :key="food.id"
+      >
+        <h3>{{ food.name }}</h3>
         <p>Data waznosci: {{ food.expirationDate }}</p>
-        <button @click="deleteFood(food.id)" class="deleteButton">Delete</button>
+        <button @click="deleteFood(food.id)" class="deleteButton">
+          Delete
+        </button>
       </li>
     </ul>
   </div>
@@ -20,10 +28,11 @@ export default {
     };
   },
   methods: {
-    calcuteDaysToExpire(food){
+    calcuteDaysToExpire(food) {
       const actualDate = new Date();
       const expireDate = new Date(food.expirationDate);
-      const difference = (expireDate.getTime() - actualDate.getTime()) / (1000 * 3600 * 24);
+      const difference =
+        (expireDate.getTime() - actualDate.getTime()) / (1000 * 3600 * 24);
       console.log(difference);
       console.log(difference <= 2);
       return difference;
@@ -45,57 +54,75 @@ export default {
               id: id,
               name: data[id].name,
               expirationDate: data[id].expirationDate,
-              difference: this.calcuteDaysToExpire(data[id])
+              difference: this.calcuteDaysToExpire(data[id]),
             });
           }
-          results.sort(function(a,b){
+          results.sort(function (a, b) {
             return new Date(a.expirationDate) - new Date(b.expirationDate);
-          })
+          });
           this.foods = results;
         });
     },
-    deleteFood(id){
-      const output = this.foods.filter((food) => food.id !== id);
-      fetch(
-        "https://food-expiration-app-default-rtdb.europe-west1.firebasedatabase.app/foods/"+ id +".json"
-      ,{
-        method: "DELETE",
-        headers: {
-          'Content-Type': 'application/json'
-        }
-      })
-      this.foods = output;
-    }
+    deleteFood(id) {
+      if (confirm("Are you sure?")) {
+        const output = this.foods.filter((food) => food.id !== id);
+        fetch(
+          "https://food-expiration-app-default-rtdb.europe-west1.firebasedatabase.app/foods/" +
+            id +
+            ".json",
+          {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+            },
+          }
+        );
+        this.foods = output;
+      }
+    },
   },
-  mounted(){
+  mounted() {
     this.loadData();
-  }
+  },
 };
 </script>
 
 <style scoped>
+ul {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+}
 .card li {
   color: white;
   background-color: rgb(80, 167, 85);
   width: 500px;
+
   margin-left: auto;
   margin-right: auto;
   border-radius: 24px;
   box-shadow: 3px 3px 3px rgba(0, 0, 0, 0.26);
-  margin-top: 30px;
-  padding: 5px 0;
-  list-style: none;
+  margin-top: 20px;
+  padding: 20px 0;
 }
-.alert{
+.alert {
   background-color: rgb(230, 175, 93) !important;
 }
-.expired{
+.expired {
   background-color: rgb(230, 93, 93) !important;
 }
-.deleteButton{
-  border: none;
+.deleteButton {
+  border-color: white;
+  border-style: double;
+  padding: 10px;
   background: none;
   color: white;
   cursor: pointer;
+}
+.deleteButton:hover {
+  filter: brightness(20%);
+}
+h3 {
+  margin-top: 0;
 }
 </style>
